header cart commit

+++ app/Livewire/Frontend/Carts/HeaderCarts.php
@@ -5,14 +5,23 @@ namespace App\Livewire\Frontend\Carts;
 use App\Enums\CartType;
 use App\Models\ProductPrice;
 use App\Models\UserCart;
+use Illuminate\Support\Facades\Session;
 use Livewire\Attributes\On;
 use Livewire\Component;
 
 class HeaderCarts extends Component
 {
+    public $total_price = 0;
+    public $discount_price = 0;
     public function deleteCart($cart_id)
     {
-        $user_cart = UserCart::query()->find($cart_id);
+        $user_cart = UserCart::query()
+            ->where('id', $cart_id)
+            ->where('user_id', auth()->id())
+            ->first();
+        if(!$user_cart){
+            return;
+        }
         $user_cart->delete();
         $this->dispatch('deleteProductCart');
     }
@@ -24,18 +33,29 @@ class HeaderCarts extends Component
     public function render()
     {
         $carts = UserCart::query()
+            ->with('product')
             ->where('user_id',auth()->id())
-            ->where('type',CartType::Main->value)->get();
+            ->where('type',CartType::Main->value)
+            ->get();
 
-        $total_price = 0;
-        foreach ($carts as $cart){
-            $product_price = ProductPrice::query()
-                ->where('product_id',$cart->product_id)
-                ->where('color_id',$cart->color_id)
-                ->where('guaranty_id',$cart->guaranty_id)
-                ->first();
-            $total_price += ($product_price->price) * $cart->count;
+        $this->total_price = 0;
+        $this->discount_price = 0;
+
+        foreach ($carts as $cart) {
+            $this->total_price += $cart->product->final_price;
+            $this->discount_price += (
+                $cart->product->main_price -
+                $cart->product->final_price
+            );
         }
-        return view('livewire.frontend.carts.header-carts',compact('carts','total_price'));
+
+        $final_price = $this->total_price;
+        $final_price = max($final_price, 0);
+        return view('livewire.frontend.carts.header-carts',[
+            'carts' => $carts,
+            'total_price' => $this->total_price,
+            'discount_price' => $this->discount_price,
+            'final_price' => max($final_price, 0),
+        ]);
     }
 }
